20.11.03 notice, category adding function complete

+++ Webmark/src/main/java/com/webmark/logic/Category.java
@@ -7,7 +7,7 @@ import com.webmark.model.CategoryVO;
 public interface Category {
 	
 	public List<CategoryVO> getList(String userid);
-	public Integer add (CategoryVO vo);
+	public Long add (CategoryVO vo);
 	public Integer delete (long cat_no);
 
 }
